Semi-Major view changes.
Working on correctly using the MVC. Going to finish tomorrow. Hard to keep working on  view without updating.
