+++ Instantiation.java
@@ -1,4 +1,5 @@
 //import Model.*;
+import Model.MySQLDatabase;
 import View.*;
 /**
  * Main method class, used to instantiate whole fleet of programs/ classes.
@@ -9,10 +10,11 @@ public class Instantiation {
         //ab1234 (al baker)
         //5f4dcc3b5aa765d61d8327deb882cf99
         //runs the program
+        MySQLDatabase msdb = new MySQLDatabase();
         new Thread() {
             @Override
-            public void run() {
-                javafx.application.Application.launch(LoginView.class);
+            public void run(){
+                javafx.application.Application.launch(MasterView.class);
             }
         }.start();
     }
